app/TradingAgents/requirements.txt
python-dotenv>=1.0.0
openai>=1.0.0
pandas-ta>=0.3.14b0
numpy>=1.24.0
pandas>=2.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
langgraph>=0.0.10
requests>=2.31.0
python-dateutil>=2.8.2
pytz>=2023.3
